# apps/backend/auth/auth-app-api/contracts/openapi/auth-app-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AuthSessionViewDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Bad Request"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [400],
                      "example": 400
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Unauthorized"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [401],
                      "example": 401
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Forbidden"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [403],
                      "example": 403
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Conflict"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [409],
                      "example": 409
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Too Many Requests"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [429],
                      "example": 429
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Internal Server Error"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [500],
                      "example": 500
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "tags": ["Auth"]
      }
    },
...
  },
  "info": {
    "title": "auth-app-api",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "nrb.sid": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nrb.sid",
        "description": "Development/default HTTP session cookie. SESSION_COOKIE_NAME may override the runtime name."
      },
      "__Host-nrb.sid": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-nrb.sid",
        "description": "Production default HTTPS session cookie. SESSION_COOKIE_NAME may override the runtime name."
      }
    },
    "schemas": {
      "AuthenticatedUserViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "displayName": {
            "type": "string"
          },
          "locale": {
            "type": "string",
            "enum": ["en", "ru"]
          },
          "theme": {
            "type": "string",
            "enum": ["system", "light", "dark"]
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri",
            "maxLength": 2048,
            "nullable": true
          },
          "avatarStatus": {
            "type": "string",
            "enum": ["none", "provider", "manual", "deleted"]
          }
        },
        "required": ["id", "tenantId", "email", "theme", "roles", "permissions"]
      },
      "AuthSessionViewDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUserViewDto"
          },
          "amr": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "authProvider": {
            "type": "string",
            "enum": ["password", "telegram", "discord"]
          },
          "authChannel": {
            "type": "string",
            "enum": ["password", "telegram_oidc", "telegram_tma", "telegram_bot", "discord_oauth", "discord_bot"]
          },
          "authTime": {
            "type": "number"
          },
          "externalIdentityId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["user"]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "writeOnly": true
          },
          "displayName": {
            "type": "string",
            "example": "Ada Lovelace"
          },
          "locale": {
            "type": "string",
            "enum": ["en", "ru"]
          }
        },
        "required": ["email", "password"]
      },
...
    }
  }
}